Add files via upload
This program checks other programs or files and returns True or False if the program/file has an even number of open and close parentheses, braces, and/or brackets. Also tracks the number of times java keywords are used and prints how many times they occur. The Test.txt file uses the original program as the input.